Remove yaml parsing code
