+++ requirements.txt
@@ -5,4 +5,4 @@ reuse
 PyYAML
 xlrd==1.2.0
 XlsxWriter==1.2.9
-fosslight_util>=1.1.0
+fosslight_util>=1.3.18
